requirements.txt
langgraph
supabase
fastapi
pydantic-settings
langchain-openai
python-dotenv
asyncpg
psycopg-binary
psycopg-pool
llama-index
llama-index-vector-stores-supabase
llama-index-embeddings-openai
pathway
pandas
pypdf
uvicorn[standard]
websockets
pyngrok
psycopg[binary]
pandas
